Removed DebugUtilityShape by merging it with DebugShapeSystem.
