+++ Havtorn/Source/Engine/ECS/ECSInclude.h
@@ -9,7 +9,7 @@
 #include "ECS/Systems/RenderSystem.h"
 #include "ECS/Systems/CameraSystem.h"
 #include "ECS/Systems/LightSystem.h"
-#include "ECS/Systems/DebugUtilitySystem.h"
+#include "ECS/Systems/DebugShapeSystem.h"
 
 #include "ECS/Components/TransformComponent.h"
 #include "ECS/Components/StaticMeshComponent.h"
