--- Havtorn/Source/Engine/ECS/Systems/DebugUtilitySystem.h
@@ -1,17 +1,0 @@
-// Copyright 2022 Team Havtorn. All Rights Reserved.
-
-#pragma once
-#include "ECS/System.h"
-
-namespace Havtorn
-{
-	class CDebugUtilitySystem final : public ISystem
-	{
-	public:
-		CDebugUtilitySystem(CScene* scene);
-		~CDebugUtilitySystem() override;
-
-		void Update(CScene* scene) override;
-
-	};
-}
